// src/nodecg/generated/speedcontrolUserAddition.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type SpeedcontrolUserAddition = {
  id: string;
  social: {
    nico?: string;
    youtube?: string;
    twitter?: string;
  };
};